+++ src/view/animation.cpp
@@ -3,17 +3,17 @@
 
 #include <SFML/Graphics/RenderWindow.hpp>
 
-void view::animation::print(sf::RenderWindow &window, int posx, int posy) const noexcept
+void view::animation::print(sf::RenderWindow &window, float posx, float posy) const noexcept
 {
 	const auto &tiles = program_state::global->resource_manager.tiles_infos();
-	auto& viewer = program_state::global->viewer;
+	auto &viewer      = program_state::global->viewer;
 
 	auto level_size = viewer.level_size();
-	auto shift = level_size.second * tiles.y_xshift;
+	auto shift      = static_cast<float>(level_size.second * tiles.y_xshift);
 
 	auto selected_frame = (viewer.current_frame() / SINGLE_IMAGE_DURATION) % m_frames.size();
 	sf::Sprite &frame   = m_frames[selected_frame];
-	frame.setPosition(static_cast<float>(posx * tiles.xspacing + posy * -tiles.y_xshift + shift),
-	                  static_cast<float>(posy * tiles.yspacing + posx * tiles.x_yshift));
+	frame.setPosition(posx * static_cast<float>(tiles.xspacing) + posy * static_cast<float>(-tiles.y_xshift) + shift,
+	                  posy * static_cast<float>(tiles.yspacing) + posx * static_cast<float>(tiles.x_yshift));
 	window.draw(frame);
 }
